修改数据库，使其符合题目要求
