--- Server/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Server
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
